Changed attributes type from float to double

+++ src/main/java/aula/lojaoculos/model/OculosGrau.java
@@ -1,11 +1,11 @@
 package aula.lojaoculos.model;
 
 public class OculosGrau extends Oculos {
-    private float grauEsquerda;
-    private float grauDireita;
+    private double grauEsquerda;
+    private double grauDireita;
     private String tipoLente;
 
-    public OculosGrau(String modelo, String marca, double preco, float grauEsquerda, float grauDireita, String tipoLente) {
+    public OculosGrau(String modelo, String marca, double preco, double grauEsquerda, double grauDireita, String tipoLente) {
         super(modelo, marca, preco);
         this.grauEsquerda = grauEsquerda;
         this.grauDireita = grauDireita;
